add FileScale codebase

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/blockmanagement/BlockManagerSafeMode.java
@@ -584,7 +584,7 @@ class BlockManagerSafeMode {
       return;
     }
 
-    int activeBlocks = blockManager.getActiveBlockCount();
+    long activeBlocks = blockManager.getActiveBlockCount();
     synchronized (this) {
       if (blockTotal != activeBlocks &&
           !(blockSafe >= 0 && blockSafe <= blockTotal)) {
